17-Jun-2015 - JWL
Updated for the following:
Port object can exist on any Block in Simulink. Changed Metamodel and
Simulink2Source rules to reflect this.
Function Block was not having a generalization connected. Removed
generation of Parent Function block. SysML2Simulink should see if a
block has a Function associated to determine type as "S-Function"